pyproject.toml
[build-system]
requires = ["setuptools>=77.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clawjournal"
version = "0.0.2"
description = "Review, score, and curate your coding agent conversation traces locally"
requires-python = ">=3.10"
license = "Apache-2.0"
license-files = ["LICENSE"]
readme = "README.md"
authors = [
    {name = "kai-rayward"},
]
keywords = ["claude-code", "codex", "gemini-cli", "opencode", "openclaw", "dataset", "conversations"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/kai-rayward/clawjournal"
Issues = "https://github.com/kai-rayward/clawjournal/issues"

[project.scripts]
clawjournal = "clawjournal.cli:main"

[tool.setuptools.packages.find]
include = ["clawjournal*"]
exclude = ["clawjournal.web*"]

[tool.setuptools.package-data]
"clawjournal" = [
    "prompts/agents/*/*.md",
    "web/frontend/dist/**/*",
    "events/docs/*.md",
    "events/docs/_features.yaml",
]

[project.optional-dependencies]
dev = ["pytest", "tomli; python_version < '3.11'"]
